Fix DrawCircleWrapper() so it takes floats instead of ints

+++ CryonicEngine/RaylibDrawWrapper.cpp
@@ -18,7 +18,7 @@ void DrawRectangleWrapper(float posX, float posY, float sizeX, float sizeY, floa
         { colorR, colorG, colorB, colorA });
 }
 
-void DrawCircleWrapper(int posX, int posY, float radius, unsigned char colorR, unsigned char colorG, unsigned char colorB, unsigned char colorA)
+void DrawCircleWrapper(float posX, float posY, float radius, unsigned char colorR, unsigned char colorG, unsigned char colorB, unsigned char colorA)
 {
     DrawCircle(posX, posY, radius, { colorR, colorG, colorB, colorA });
 }
